MapUns_api/requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
reportlab==3.4.0
xlwt==1.3.0
